% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/team_6.R
\name{team_6}
\alias{team_6}
\title{Function from Team 6}
\usage{
team_6(file, tolerance)
}
\arguments{
\item{file}{A path having shape file of selected country.}

\item{tolerance}{A number for tolerence.}
}
\value{
The dataframe about how to plot the map of a country such as Australia as default
examples
file<-"gadm36_AUS_shp/gadm36_AUS_1.shp"
tolerance <- 0.1
team_6(file="gadm36_AUS_1.shp",0.01)
team_6(file,0.1) %>% ggplot(aes(x = long, y = lat, group = group)) + geom_polygon(fill="white",color="black",lwd=1)
}
\description{
Function from Team 6
}
