% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/team_7.R
\name{team_7}
\alias{team_7}
\title{Solution of team 7}
\usage{
team_7(file, tolerance)
}
\arguments{
\item{file}{A path of the file}

\item{tolerance}{A number}
}
\value{
A dataframe of the geographic information of the polygons
examples
team_7(file="gadm36_AUS_1.shp",0.01)
}
\description{
Solution of team 7
}
